chore: Dependency fixes (remove cross-dep to apps/web)
