+++ packages/features/apps/components/AppListCard.tsx
@@ -3,9 +3,33 @@
 import classNames from "@calcom/lib/classNames";
 import { getPlaceholderAvatar } from "@calcom/lib/defaultAvatarImage";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
-import { Avatar, Badge, Icon, ListItemText } from "@calcom/ui";
+import { Avatar } from "@calcom/ui/avatar";
+import { Badge } from "@calcom/ui/badge";
+import { Icon } from "@calcom/ui/icon";
+import { ListItemText } from "@calcom/ui/list";
 
-import type { AppListCardProps } from "../../../../apps/web/components/AppListCard";
+type ShouldHighlight =
+  | {
+      slug: string;
+      shouldHighlight: true;
+    }
+  | {
+      shouldHighlight?: never;
+      slug?: never;
+    };
+
+export type AppListCardProps = {
+  logo?: string;
+  title: string;
+  description: string;
+  actions?: ReactNode;
+  isDefault?: boolean;
+  isTemplate?: boolean;
+  invalidCredential?: boolean;
+  children?: ReactNode;
+  credentialOwner?: CredentialOwner;
+  className?: string;
+} & ShouldHighlight;
 
 export const AppListCard = (props: AppListCardProps & { highlight?: boolean }) => {
   const { t } = useLocale();
