--- packages/ui/components/app-list-card/index.ts
@@ -1,1 +1,0 @@
-export { AppListCard } from "./AppListCard";
