+++ packages/ui/components/apps/AppCard.tsx
@@ -14,10 +14,11 @@ import { useLocale } from "@calcom/lib/hooks/useLocale";
 import type { UserAdminTeams } from "@calcom/lib/server/repository/user";
 import type { AppFrontendPayload as App } from "@calcom/types/App";
 import type { CredentialFrontendPayload as Credential } from "@calcom/types/Credential";
-import type { ButtonProps } from "@calcom/ui";
-import { Badge, showToast } from "@calcom/ui";
 
+import { Badge } from "../badge";
+import type { ButtonProps } from "../button";
 import { Button } from "../button";
+import { showToast } from "../toast";
 
 interface AppCardProps {
   app: App;
