+++ packages/ui/components/avatar/UserAvatar.tsx
@@ -3,7 +3,8 @@ import { getPlaceholderAvatar } from "@calcom/lib/defaultAvatarImage";
 import { getUserAvatarUrl } from "@calcom/lib/getAvatarUrl";
 import type { User } from "@calcom/prisma/client";
 import type { UserProfile } from "@calcom/types/UserProfile";
-import { Avatar } from "@calcom/ui";
+
+import { Avatar } from "./Avatar";
 
 type Organization = {
   id: number;
